+++ src/__playground__/worker.ts
@@ -1,5 +1,14 @@
-import { define } from 'nanolith';
+import { ParentThread, define } from 'nanolith';
 
 export const worker = await define({
-    yourTask() {},
+    __initializeService() {
+        console.log('throwing error');
+        throw new Error();
+    },
+    async task() {
+        while (true) {
+            await new Promise((r) => setTimeout(r, 1e3));
+            ParentThread.sendMessage('foo');
+        }
+    },
 });